refactor(structure): Organize .html and .js files, add styles to pages

+++ assets/js/delete_incident.js
@@ -13,10 +13,11 @@ function createElements(){
     let delete_button = document.createElement("button");
     delete_button.id = "delete_button";
     delete_button.textContent = "Eliminar";
+    delete_button.type = "button";
 
-    document.appendChild(title);
-    document.appendChild(input_id);
-    document.appendChild(delete_button); 
+    document.body.appendChild(title);
+    document.body.appendChild(input_id);
+    document.body.appendChild(delete_button); 
 }
 
 function addStyle(){
@@ -24,6 +25,8 @@ function addStyle(){
 }
 
 function deleteIncident(id){
+    if (!confirm(`¿Estás seguro de que querés eliminar el incidente con ID ${id}?`)) return;
+
     fetch(`http://localhost:3000/incidents/${id}`, {
         method: "DELETE",
         headers: {
@@ -34,9 +37,11 @@ function deleteIncident(id){
     .then(response => response.json())
     .then(data => {
         console.log("Incidente eliminado", data);
+        alert("Incidente eliminado");
     })
     .catch(error => {
         console.error("Error al eliminar incidente", error);
+        alert("Error al eliminar incidente");
     });
 }
 
@@ -48,7 +53,17 @@ function main (){
     let input_id = document.getElementById("input_id");
     let button = document.getElementById("delete_button"); 
 
-    button.addEventListener("click", () => 
-        deleteIncident(input_id.value)
-    );
+
+    button.addEventListener("click", () => {
+        const id = input_id.value.trim();
+
+        if (!id) {
+            alert("Por favor ingrese un ID válido.");
+            return;
+        }
+        deleteIncident(id);
+        input_id = "";
+    });
 }
+
+main();
